Allow passing in AWS creds when initialising
We need to use environment variables that end in _ACCESS_KEY_ID and
_SECRET_ACCESS_KEY because Doppler says so.

In my opinion the better way to deal with that is to allow us to pass in
authorisation when initialising the client. This means that the system
is in control of what it names the environment variables.